pyproject.toml
[project]
requires-python = ">=3.11"
name = "BCL-Examples"
version="1.0.0"
dependencies=["requests"]

[project.optional-dependencies]
google = ["google-auth", "google-oauth"]
